Add ability to store up to 10 colors in local storage upon picking them.
