+++ src/picker/index.js
@@ -8,6 +8,7 @@ import {
     getCustomColorCache,
     setCustomColorCache
 } from './utils/main';
+import LocalStorageService from "../localStorageService";
 const ColorCollections = ['#ff1300','#EC7878','#9C27B0','#673AB7','#3F51B5','#0070FF','#03A9F4','#00BCD4','#4CAF50','#8BC34A','#CDDC39','#FFE500','#FFBF00','#FF9800','#795548','#9E9E9E','#5A5A5A','#FFF'];
 class ColorPlugin extends HTMLElement {
 
@@ -229,6 +230,7 @@ class ColorPlugin extends HTMLElement {
                 this.nativeclick = true;
                 this.value = handleCSSVariables(ev.target.value);
                 this.onColorPicked(this.value);
+                LocalStorageService.setColorInLocalStorage(this.value)
                 setCustomColorCache(this.value, this.pluginType);
 
                 customPicker.style.backgroundColor = this.value;
